server/requirements.txt
websockets>=12.0,<13.0
asyncio>=3.4.3
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
aioesphomeapi>=24.6.0
noise>=1.2.2
protobuf>=4.21.0
zeroconf>=0.131.0
aiohttp
